+++ app/controllers/application_controller.rb
@@ -1,6 +1,7 @@
 class ApplicationController < ActionController::API
   # protect_from_forgery with: :null_session
   before_action :authenticate_request
+  before_action :check_meets
 
   private
 
@@ -20,5 +21,10 @@ class ApplicationController < ActionController::API
 
   def current_user
     @current_user
+  end
+
+  # Check if all the meeting been in the past are finished
+  def check_meets
+    MeetsService.date_check()
   end
 end