+++ app/services/meets_service.rb
@@ -1,0 +1,5 @@
+class MeetsService
+    def self.date_check
+        Meet.mark_finished_meets
+    end
+end